backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
python-dotenv
psycopg2-binary
Flask-Migrate